// apps/mobile/lib/shared/integration-types.ts
export type SourceModule =
  | 'PC-1'
  | 'PC-2'
  | 'S-1'
  | 'S-2'
  | 'C-1'
  | 'C-2'
  | 'OH-1'
  | 'CIE-1'
  | 'CIE-2'
  | 'CIE-3'
  | 'CIE-4';